DBT/Snowflake/snowflake REST API/requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
atpublic==5.0
azure-core==1.31.0
azure-identity==1.19.0
azure-keyvault-secrets==4.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
distlib==0.3.8
Faker==30.3.0
filelock==3.16.1
idna==3.10
isodate==0.7.2
msal==1.31.0
msal-extensions==1.2.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
portalocker==2.10.1
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
snowflake-connector-python==3.12.1
snowflake._legacy==0.13.0
snowflake.core==0.13.0
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
